requirements.txt
# InvestingDojo Automation Requirements
# Install with: pip install -r automation/requirements.txt

# RSS Feed Processing
feedparser==6.0.10

# HTTP Requests
requests==2.31.0

# Audio Processing
pydub==0.25.1

# Google Gemini AI
google-generativeai==0.3.2

# Database
mysql-connector-python==8.2.0

# Utilities
python-dateutil==2.8.2
pathlib2==2.3.7

# CrewAI Framework Dependencies
crewai==0.150.0
crewai-tools==0.58.0

# CrewAI Additional Dependencies
pydantic>=2.5.0
langchain>=0.1.0
langchain-community>=0.0.13
langchain-core>=0.1.12
openai>=1.12.0

# Optional: For better audio processing
# ffmpeg-python==0.2.0  # Requires ffmpeg installed separately